My last commit wouldn't compile. This version compiles but has some old code.
I don't know what happened last night but the stash we made was created after some stuff already got fucked up with some old code. I unlinked the bpmLabel since it was linked to something that didn't exist anymore, and defined the DATA_RATE constant in WWHeartRateDetector.swift since a part of the code was relying on it and couldn't find it.