Switched location data to Mapbox's built in method
Deleted my Core Location code and switched to Mapox's built-in method for location tracking. Although it might give up too much control over accuracy, I'd like to think that it will work, as it really cleans up the code.

Also added an app icon and switched from creating the mapview in code to creating it in the storyboard.